requirements.txt
torch
torchvision
torchaudio
gradio
transformers
accelerate
tokenizers
datasets
Pillow
numpy
requests
scipy
timm